v0.1.8.2: refactoring & fixes
