+++ src/hooks/index.ts
@@ -1,2 +1,2 @@
 export * from './ui'
-export * from './web3'
+export * from './useMarketData'
