--- src/hooks/web3/index.ts
@@ -1,1 +1,0 @@
-export * from './useMarketData'
