+++ src/providers/Wagmi.tsx
@@ -2,7 +2,7 @@
 // import { publicProvider } from 'wagmi/providers/public'
 // import { jsonRpcProvider } from 'wagmi/providers/jsonRpc'
 import { defaultChain } from '@/constants'
-import { web3AuthConnector } from '@/libs'
+import { web3AuthConnector } from '@/providers'
 import { createPublicClient } from 'viem'
 import { WagmiProvider as WagmiDefaultProvider, http, createConfig } from 'wagmi'
