+++ src/providers/index.ts
@@ -1,6 +1,4 @@
 export * from './Web3Auth'
 export * from './Wagmi'
-export * from './Amplitude'
-export * from './Etherspot'
 export * from './ReactQuery'
 export * from './Chakra'
