+++ src/services/BalanceService.tsx
@@ -2,13 +2,14 @@ import { Toast, ToastWithdraw } from '@/components'
 import { collateralToken, defaultChain } from '@/constants'
 import { erc20ABI } from '@/contracts'
 import { useToast } from '@/hooks'
-import { publicClient, useEtherspot } from '@/libs'
+import { publicClient } from '@/providers'
+import { useEtherspot } from '@/services'
 import { Address, GetBalanceResult } from '@/types'
-import { Logger, NumberUtil } from '@/utils'
+import { Logger } from '@/utils'
 import { useMutation, useQuery } from '@tanstack/react-query'
 import { PropsWithChildren, createContext, useContext, useEffect, useMemo, useState } from 'react'
 import { formatUnits, getContract, isAddress, parseUnits } from 'viem'
-import { useAccount, useChainId } from 'wagmi'
+import { useAccount } from 'wagmi'
 
 interface IBalanceService {
   strategy: 'Deposit' | 'Withdraw'
