+++ src/services/TradingService.tsx
@@ -2,8 +2,8 @@ import { Toast } from '@/components'
 import { collateralToken, defaultChain } from '@/constants'
 import { marketMakerABI } from '@/contracts'
 import { useToast } from '@/hooks'
-import { publicClient, useEtherspot } from '@/libs'
-import { useAccount, useBalanceService, useHistory } from '@/services'
+import { publicClient } from '@/providers'
+import { useAccount, useBalanceService, useEtherspot, useHistory } from '@/services'
 import { Market } from '@/types'
 import { NumberUtil } from '@/utils'
 import { sleep } from '@etherspot/prime-sdk/dist/sdk/common'
