+++ src/services/index.ts
@@ -1,4 +1,6 @@
 export * from './LimitlessApi'
+export * from './Etherspot'
+export * from './Amplitude'
 export * from './AccountService'
 export * from './BalanceService'
 export * from './TradingService'
